Enable gzip, convert gif files, split sudokuboard out, better acknowledgments
