+++ Dockerfile
@@ -1,6 +1,8 @@
 FROM jamesandersen/alpine-golang-opencv3:edge
+RUN apk --no-cache add git
 WORKDIR /go/src/github.com/jamesandersen/gosudoku
 COPY . .
+RUN go get github.com/nytimes/gziphandler
 RUN go-wrapper download   # "go get -d -v ./..."
 RUN go-wrapper install    # "go install -v ./..."
