+++ web/sudoku.html
@@ -1,6 +1,5 @@
 <!DOCTYPE html>
 <html>
-
 <head>
     <title>OpenCV Sudoku Solver</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
@@ -46,11 +45,41 @@
     <div id="ack-row" class="acknowledgements">
             <h3 id="ack"><span class="arrow">⬆</span>&nbsp;Acknowledgements</h3>
             <ul>
-                <li><a href="http://docs.opencv.org/3.2.0/">OpenCV 3.2</a> - Parsing the Sudoku image and the SVM machine learning model to detect digits rely on OpenCV 3.</li>
-                <li><a href="https://www.udacity.com/ai">Udacity AI Nanodegree</a> - The Sudoku solving code was written originally as assignment in Udacity's fine AI Nanodegree program.  For this project I ported it to <a href="https://golang.org/">Go</a>.</li>
-                <li><a href="https://github.com/edankwan/PerspectiveTransform.js">PerspectiveTransform.js</a> - calculating the appropriate CSS matrix3d transform to apply the solution back on the image.</li>
-                <li><a href="https://projects.lukehaas.me/css-loaders/">CSS Loaders</a> - Luke Haas for his cool CSS loaders</li>
-                <li><a href="https://hyper.sh/">Hyper.sh</a> - Hosting of the <a href="https://hub.docker.com/r/jamesandersen/gosudoku/">custom docker container</a> running this app is provided by Hyper.sh</li>
+                <li>
+                    <a class="logo" href="http://docs.opencv.org/3.2.0/">
+                        <img src="/static/icons/opencv.png" alt="OpenCV logo"/>
+                        <div class="title">OpenCV 3.2.0</div>
+                    </a>
+                    <div class="desc">Provides routines cobbled together to parse the Sudoku image and the SVM machine learning model to detect digits</div>
+                </li>
+                <li>
+                    <a class="logo" href="https://www.udacity.com/ai">
+                        <img src="/static/icons/udacity.png" alt="Udacity logo"/>
+                        <div class="title">Udacity AI Nanodegree</div>
+                    </a>
+                    <div class="desc">The Sudoku solving code was written originally as assignment in Udacity's fine AI Nanodegree program.  For this project I ported it to <a href="https://golang.org/">Go</a>.</div>
+                </li>
+                <li>
+                    <a class="logo" href="https://github.com/edankwan/PerspectiveTransform.js">
+                        <img src="/static/icons/github.png" alt="Github PerspectiveTransform.js"/>
+                        <div class="title">PerspectiveTransform.js</div>
+                    </a>
+                    <div class="desc">Calculates the appropriate CSS matrix3d transform to apply the solution back on the puzzle image</div>
+                </li>
+                <li>
+                    <a class="logo" href="https://projects.lukehaas.me/css-loaders/">
+                        <img src="/static/icons/github.png" alt="CSS Loaders"/>
+                        <div class="title">CSS Spinners</div>
+                    </a>
+                    <div class="desc"><a href="https://twitter.com/lukehaas">Luke Haas</a> gets credit for his slick CSS loader</div>
+                </li>
+                <li>
+                    <a class="logo" href="https://hyper.sh/">
+                        <img src="/static/icons/hyper.jpg" alt="CSS Loaders"/>
+                        <div class="title">Hyper.sh</div>
+                    </a>
+                    <div class="desc">Hosts of the <a href="https://hub.docker.com/r/jamesandersen/gosudoku/">custom docker container</a> running this app is provided by Hyper.sh</div>
+                </li>
             </ul>
         </div> 
     <a href="https://github.com/jamesandersen/go-sudoku"><img class="github" src="https://camo.githubusercontent.com/e7bbb0521b397edbd5fe43e7f760759336b5e05f/68747470733a2f2f73332e616d617a6f6e6177732e636f6d2f6769746875622f726962626f6e732f666f726b6d655f72696768745f677265656e5f3030373230302e706e67" alt="Fork me on GitHub" data-canonical-src="https://s3.amazonaws.com/github/ribbons/forkme_right_green_007200.png"></a>
